Hoist trust-boundary envelope primitives into core/envelope.js
The reserved-envelope-key registry and strip walkers
(RESERVED_ARG_KEYS, RESERVED_ENVELOPE_KEYS, EGRESS_STRIP_KEYS,
MAX_STRIP_DEPTH, stripInternalEnvelopes, stripEgressEnvelopes,
sanitizeResultObject, sanitizeTransportEgress) lived in bridge.js, and
mixer, connect, webhook, and reverse all reached into the largest
orchestrator file for them. connect.js additionally carried its own
copy of the strip walker just to add a security log line.

Move them to src/core/envelope.js as the single registry. bridge.js
re-exports every public name, so existing imports (invariant tests,
red-team suites, stress scripts) are unchanged. connect.js's
stripUpstreamEnvelopes becomes a thin wrapper over the shared walker
via a new onStrip callback, preserving its per-key SECURITY stderr
line. SPEC §2 now points at core/envelope.js as the registry: adding a
reserved key there is sufficient for every dispatch surface.

No behavior change. Lint clean, 1472 tests pass / 0 fail (5 network
skips), trust-matrix 11/11.

Closes #44
