+++ src/compose/mixer.js
@@ -11,7 +11,8 @@ import { createStdioTransport } from '../transport/stdio.js';
 import { applyResponseTransform } from '../transforms/response.js';
 import { executeChain } from './chain.js';
 import { resolveEnvVars } from '../core/env.js';
-import { validateToolArgs, sanitizeTransportEgress, emitAuditLog } from '../bridge.js';
+import { validateToolArgs, emitAuditLog } from '../bridge.js';
+import { sanitizeTransportEgress } from '../core/envelope.js';
 import { sanitizeMcpToolDescription } from '../core/sanitize.js';
 import { safeLog } from '../core/events.js';
 import '../core/types.js';
